backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.10.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
# NLP and ML dependencies
spacy>=3.7.0
transformers>=4.35.0
sentence-transformers>=2.2.0
nltk>=3.8.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
# Google Gemini Pro API
google-generativeai>=0.3.0

